% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print "fars_map_state"}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric vector, specifying the state which should be plotted.}

\item{year}{a numeric vector of length one, specifying for which year the accidents should be plotted.}
}
\value{
This function has no return value. It generates a plot with all accidents of the corresponding year and state.
}
\description{
This function plot a map where all accidents for the specified states (\code{state.num})  and the specified \code{year} are mapped, dependent on a csv file where these data are stored.
}
\details{
This function incluede the \link{readr::read_csv}, \link{dplyr::tbl_df}, \link{dplyr::filter} and \link{maps::map}  functions.
         Therefore the dplyr, maps and readr package has to be installed and loaded.
}
\note{
If for the year no file exists an error occurs. If the state specifyed with \code{state.num} does not exist in the data an error occors. 
      If no coordinates exists for any of the accidents an error occurs.
}
\examples{
\dontrun{fars_map_state(01 , 2013)}
\dontrun{fars_map_state(01 , 2014)}
\dontrun{fars_map_state(02 , 2014)}
}
