% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_functions.R
\name{fars_read}
\alias{fars_read}
\title{"fars_read"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character vector of length one which specifies the data path and/or filename which should be loaded.}
}
\value{
This function returns a tibble with the contents of the csv. file
}
\description{
This is a function that loads a csv file (specified using the \code{filename} argument) and returns it as tibble.
}
\details{
This function is a wrapper for the \code{\link{readr::read_csv}} and \code{\link{dplyr::tbl_df}} functions. Therefore
         the dplyr and readr package has to be installed and loaded.
}
\note{
If the \code{filename} argument does not represent a file an error occurs.
}
\examples{
\dontrun{ mydata <- fars_read("accident_2013.csv")}
\dontrun{myData <- fars_read("accident_2014.csv")}
}
