% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{"fars_read_years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a numeric vector, representing the years for which the accident_data should be loaded.}
}
\value{
This function returns list for all the years specified in the \code{years} argument. Every entry of the list is a data.frame with the columns year and MONTH.
}
\description{
This is a function that loads the columns MONTH of the accident data for all specified years (specified using the \code{years} argument) and returns a list.
}
\details{
This function combines the fars_read and make_filename function. The external functions \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{tbl_df}} and \code{\link[readr]{read_csv}} are included.
         Therefore the dplyr and readr package has to be installed and loaded.
}
\note{
If for an element of the \code{years} argument no file is found a warning occurs.
}
\examples{
\dontrun{myData <- fars_read_years(c(2013,2014))}
\dontrun{myData <- fars_read_years(2014)}
}
