% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcacalc.R
\name{input data}
\alias{input data}
\alias{pcabrca}
\title{pcabrca}
\usage{
pcabrca(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plots and data frames
}
\description{
pca analysis
}
\details{
this program takes input as data frame and performs principal component analysis on the data.
}
\author{
Anushka Jain
}
