% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_functions.R
\name{make_filename}
\alias{make_filename}
\title{"make_filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numerical vector, which contains the years for which a filename should be generated.}
}
\value{
This function returns a character vector containing the filenames corresponding to the  \code{year} argument.
}
\description{
This is a function that creates form the \code{year} argument a filename the corresponding accident csv.
}
\details{
This function operates on strings.
}
\note{
If the \code{year} argument is not a numerical vector the generated filename probobly do not exist.
}
\examples{
\dontrun{ make_filename(2013)}
\dontrun{make_filename(2014)}
}
