% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{inputData}
\usage{
readdata(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
read dataset
}
\details{
this program reads datasets in from the local system.
}
\author{
Anushka Jain
}
