% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{""}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector, representing the years for which the accident_data should be summarized.
fars_summarize_years}
}
\value{
This function returns a tibble which contains in the first collumn the month and in the following columns for every year in \code{years} the number of accidents. for all the years specified in the \code{years} argument. Every entry of the list is a data.frame with the columns year and MONTH.
}
\description{
This is a function summarizes the number of accidents for every month of all selected years (using the \code{years} argument) and returns tibble with the columns MONTH and for every year in \code{years} a column with the number of accidents.
}
\details{
This function is a wrapper for the \link{fars_read_years} function. The external functions \link{dplyr::mutate}, \link{dplyr::select}, \link{dplyr::group_by}, \link{dplyr::bind_rows},
          \link{dplyr::summarize}, \link{tidyr::spread} \link{dplyr::tbl_df} and \link{readr::read_csv} are included.
         Therefore the dplyr , tidyr and readr package has to be installed and loaded.
}
\note{
If for an element of the \code{years} argument no corresponding file is found a warning occurs.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))

}
